// frontend/src/api-types.ts
export interface components {
  schemas: {
    /** Author */
    Author: {
      /** Name */
      name: string;
      /** Role */
      role: string;
    };
    /** Genre */
    Genre: {
      /** Name */
      name: string;
    };
    /** HTTPValidationError */
    HTTPValidationError: {
      /** Detail */
      detail?: components["schemas"]["ValidationError"][];
    };
    /** Manga */
    Manga: {
      /** Title */
      title: string;
      /** Description */
      description: string;
      /** Total Volumes */
      total_volumes: number;
      /** Volumes */
      volumes: number[];
      /** Authors */
      authors: components["schemas"]["Author"][];
      /** Genres */
      genres: components["schemas"]["Genre"][];
    };
    /** ValidationError */
    ValidationError: {
      /** Location */
      loc: (string | number)[];
      /** Message */
      msg: string;
      /** Error Type */
      type: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}